Adds intrinsics for arm64.
Also enables tests that no longer time-out.

R=regis@google.com

Review URL: https://codereview.chromium.org//285403004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@36330 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/simulator_arm64.cc
@@ -611,7 +611,7 @@ int32_t Simulator::get_wregister(Register reg, R31Type r31t) const {
   if ((reg == R31) && (r31t == R31IsZR)) {
     return 0;
   } else {
-    return registers_[reg];
+    return static_cast<int32_t>(registers_[reg]);
   }
 }
